+++ src/resources/config/box.php
@@ -8,10 +8,18 @@ return [
         'namespace' => 'vanilo'
     ],
     'routes' => [
-        'prefix' => 'admin',
-        'as' => 'vanilo.admin.',
-        'middleware' => ['web', 'auth', 'acl'],
-        'files' => ['admin']
+        [
+            'prefix' => 'admin',
+            'as' => 'vanilo.admin.',
+            'middleware' => ['web', 'auth', 'acl'],
+            'files' => ['admin'],
+        ],
+        [
+            'prefix' => 'admin',
+            'as' => 'vanilo.admin.alt.',
+            'middleware' => ['web', 'auth'],
+            'files' => ['nonacl'],
+        ],
     ],
     'breadcrumbs' => true,
 ];