+++ Changelog.md
@@ -35,6 +35,7 @@
 - Added the `domain` field to the channel form
 - Added billing and shipping zones to the channel form
 - Added zone support to payment methods
+- Added an alternative route `vanilo.admin.alt.master_product_variant.show` that shows a variant, but doesn't require passing the master id
 
 ## 3.7.1
 ##### 2023-12-17